debian/*.service: overhaul to use guess-dsp.sh
guess-dsp.sh comes from
https://github.com/qualcomm-linux/meta-qcom/blob/9d2d2368a90a9e4faa772a10808fe4c284346e20/recipes-support/fastrpc/fastrpc/guess-dsp.sh.

I've modified guess-dsp.sh to drop mount support as I consider it
unsuitable for a service start in Debian to come with a side-effect of
mounting something. I'm told that mount support is likely not needed in
our Debian use case (else I'll implement that somewhere else).

I've also dropped the shebang, as it is not useful to execute
guess-dsp.sh as a script. It must be sourced. Otherwise Debian's lintian
correctly flags it as not executable.

I'm using Type=exec as recommended by systemd.service(5) (a richer type
like notify isn't currently supported upstream).

Signed-off-by: Robie Basak <robie.basak@oss.qualcomm.com>